Relocated Multiblock class
