+++ src/main/java/io/github/thebusybiscuit/slimefun4/core/multiblocks/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * This package holds all core mechanics related to a
+ * {@link io.github.thebusybiscuit.slimefun4.core.multiblocks.MultiBlock}, like that class itself.
+ */
+package io.github.thebusybiscuit.slimefun4.core.multiblocks;
