+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * This package holds all classes that are related to the actual implementation of this plugin.
+ * This includes implementations of {@link me.mrCookieSlime.Slimefun.Objects.SlimefunItem.SlimefunItem} but also any
+ * {@link org.bukkit.event.Listener}.
+ */
+package io.github.thebusybiscuit.slimefun4.implementation;
